FIX: don't check if user exists when local user is null

+++ src/context/UserContext.jsx
@@ -27,6 +27,12 @@ const UserProvider = ({ children }) => {
     const state = { user, setUser }
 
     useEffect(() => {
+
+        // Dont check current user if null
+        if (!user)
+            return
+
+        // Clear localStorage and user if it does not exist in the API
         checkForUser(storageRead(STORAGE_KEY_USER)?.username).then(([error, user]) => {
             if (error || user.length === 0) {
                 localStorage.clear()
